+++ src/routes/auth.js
@@ -1,15 +1,13 @@
 const Router = require('express-promise-router');
 
-const auth = require('../middleware/auth');
+const { login, signup } = require('controllers/auth');
 
 const router = Router({ mergeParams: true });
-const { login, signup, putUser } = require('../controller/user');
 
 /**
- * User routes.
+ * Auth routes.
  */
 router.post('/login', login);
 router.post('/signup', signup);
-router.patch('/user', auth, putUser);
 
 module.exports = router;
